add finalizacao de compra e exibicao dos pedidos

+++ src/main/java/com/fatec/livrariaonlinejpa/model/Pagamento.java
@@ -11,5 +11,5 @@ public class Pagamento {
     private long id;
     @ManyToOne
     private Cartao cartao;
-    private float valor;
+    private double valor;
 }
